# hummingbird/docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Hummingbird Media Processing - OpenAPI 3.1
  description: |-
    This API is part of a larger project, developed by Weder Ribas during his work in the CS7990 class (Master Thesis) at Northeast University Vancouver.
    The Hummingbird API provides a simple application that provides users with the ability to upload an image for processing, obtain metadata from this image, and download the processed image.

    Some useful links:
    - [The thesis code repository](https://github.com/wederribas/cs7990-master-thesis/tree/main)

  contact:
    email: contact@wederribas.com
  license:
    name: GPL-3.0 license
    url: https://www.gnu.org/licenses/gpl-3.0.en.html$0
  version: 0.0.1
servers:
  - url: http://hummingbird-alb.elb.localhost.localstack.cloud:4566/v1
...
paths:
  /media/upload:
    post:
      tags:
        - media
      summary: Upload a media file for processing
      description: Upload a media file for processing. The file will be processed asynchronously and the response will contain the media ID. Use the media ID to check the processing status in the `/media/{mediaId}/status` endpoint.
      operationId: uploadMedia
      requestBody:
        description: Upload a media file for processing
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/MediaUpload'
        required: true
      responses:
        '202':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadedResponse'
        '400':
          description: Invalid input
        '413':
          description: Request entity too large
        '422':
          description: Validation exception
        '500':
          description: Internal server error
...
components:
  schemas:
    MediaUploadedResponse:
      type: object
      properties:
        mediaId:
          type: string
          format: uuid
          examples: [123e4567-e89b-12d3-a456-426614174000]
    MediaUpload:
      type: object
      properties:
        media:
          type: string
          format: binary
      required:
        - media
...